done crud fix navigation stack errors

+++ src/pages/user-pop-over/user-pop-over.ts
@@ -1,7 +1,8 @@
+import { HomePage } from '../home/home';
 import { UserPage } from '../user/user';
 import { UserSettings } from '../../shared/shared';
 import { Component } from '@angular/core';
-import { Events, IonicPage, NavController, NavParams, ViewController } from 'ionic-angular';
+import { AlertController, Events, IonicPage, NavController, NavParams, ViewController } from 'ionic-angular';
 
 /**
  * Generated class for the UserPopOverPage page.
@@ -17,31 +18,55 @@ import { Events, IonicPage, NavController, NavParams, ViewController } from 'ion
 })
 export class UserPopOverPage {
 
+  userDetailsSubscribe: (res:any) => void;
+
+  user;
+
   constructor(public navCtrl: NavController, 
               public navParams: NavParams,
               private viewController: ViewController,
               private userSettings: UserSettings,
-              private events: Events) {}
+              private events: Events,
+              private alertCtrl: AlertController) {}
 
   toggleDelete(){
+    
+    let confirm = this.alertCtrl.create({
+      title: 'Delete Contact',
+      message: 'Are you sure you want to delete?',
+      buttons: [
+        {
+          text: 'Ok',
+          handler: () => {
+            this.userSettings.deleteUser(this.navParams.data.rowid)
+            .then( () => {
+              
+              this.events.publish('user:popover');
 
-    this.userSettings.deleteUser(this.navParams.data.rowid)
-      .then( () => {
-        this.viewController.dismiss();
-        this.events.publish('close:userPopOver');
-      });
+            });
+          }
+        },
+        {
+          text: 'Cancel',
+          handler: () => { /* */ }
+        }
+      ]
+    });
+
+    this.viewController.dismiss();
+    confirm.present();
   }
 
   toggleEdit(){
-
     this.viewController.dismiss();
-
-    this.navCtrl.push(UserPage, this.navParams.data);
-    
-    // this.events.publish('user:edit');
+    this.navCtrl.push(UserPage, this.user);
   }
-  
+
   ionViewDidLoad() {
     console.log('ionViewDidLoad UserPopOverPage');
+    this.user = this.navParams.data;
   }
+
+  ionViewWillLeave(){ /* */ }
+
 }
